requirements.txt
asgiref==3.6.0
charset-normalizer==3.1.0
Django==4.2.1
django-ckeditor==6.6.1
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-modeltranslation==0.18.11
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-recaptcha==3.0.0
django-taggit==4.0.0
idna==3.4
phonenumbers==8.13.13
phonenumberslite==8.13.13
Pillow==9.5.0
psycopg2==2.9.6
pymemcache==4.0.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
username==2.1.0
